Refactored java exceptions. Removed unneeded classes.
